requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-arm64
asgiref=3.7.2=pyhd8ed1ab_0
backports.zoneinfo=0.2.1=py310hbe9552e_7
brotlipy=0.7.0=py310h1a28f6b_1002
bzip2=1.0.8=h3422bc3_4
ca-certificates=2022.4.26=hca03da5_0
certifi=2022.6.15=py310hca03da5_0
cffi=1.15.0=py310h22df2f2_1
charset-normalizer=2.0.4=pyhd3eb1b0_0
click=8.1.6=unix_pyh707e725_0
cryptography=41.0.3=py310hdd3b5e7_0
django=4.2.3=pyhd8ed1ab_0
django-cors-headers=4.2.0=pyhd8ed1ab_0
djangorestframework=3.14.0=pyhd8ed1ab_0
idna=3.3=pyhd3eb1b0_0
krb5=1.21.2=h92f50d5_0
libcxx=16.0.6=h4653b0c_0
libedit=3.1.20210910=h1a28f6b_0
libffi=3.4.2=h3422bc3_5
libpq=15.4=hcea71ed_0
libsqlite=3.42.0=hb31c410_0
libzlib=1.2.13=h53f4e23_5
ncurses=6.4=h7ea286d_0
openssl=3.1.2=h53f4e23_0
pip=23.2.1=pyhd8ed1ab_0
psycopg=3.1.10=py310h4715074_0
psycopg-c=3.1.10=py310hce2aa25_0
psycopg2=2.9.6=py310h2b81375_0
pycparser=2.21=pyhd3eb1b0_0
pyopenssl=23.2.0=pyhd8ed1ab_1
pysocks=1.7.1=py310hca03da5_0
python=3.10.12=h01493a6_0_cpython
python-dotenv=1.0.0=pyhd8ed1ab_0
python_abi=3.10=3_cp310
pytz=2023.3=pyhd8ed1ab_0
readline=8.2=h92ec313_1
requests=2.27.1=pyhd3eb1b0_0
setuptools=68.0.0=pyhd8ed1ab_0
sqlparse=0.4.4=pyhd8ed1ab_0
tk=8.6.12=he1e0b03_0
typing-extensions=4.7.1=hd8ed1ab_0
typing_extensions=4.7.1=pyha770c72_0
tzdata=2023c=h71feb2d_0
urllib3=1.26.9=py310hca03da5_0
wheel=0.41.1=pyhd8ed1ab_0
xz=5.2.6=h57fd34a_0
